Removed breaking changes over a month old
The goal of this project is to inform people of upcoming breaking changes or recent ones that may have impacted their environment. As such I'll be regularly clearing out the issues that are over a month old. I suspect someone might argue there should be a historical list, and for that there is the commit log, as I'm not interested in maintaining something else for it.