Add AWS GovCloud (US) RDS CA Certificate Rotation
The AWS GovCloud RDS CA Certificates expire on May 18. In December, AWS
sent emails to customers using RDS in this partition informing them of
the coming change. There don't seem to be the same volume of
documentation and blog posts that were done when this change was made in
the commercial partition. The change is also referenced on the AWS
GovCloud (US) documentation for Amazon RDS but only as a way it differs
from the other regions.

If there's a good way to do so, I am happy to provide the content of the
email from AWS.